ai_workflow_integration_demo/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.43.3
torch>=2.1.0
shap==0.46.0
httpx==0.27.2
pytest==8.3.2
